@@ -2,42 +2,28 @@ source 'https://rubygems.org'
 
 
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.2.5.1'
-# Use postgresql as the database for Active Record
-gem 'pg', '~> 0.15'
+gem 'rails', '5.2.4.2'
 
-# Use SCSS for stylesheets
-gem 'sass-rails', '~> 5.0'
-# Use Uglifier as compressor for JavaScript assets
-gem 'uglifier', '>= 1.3.0'
-# Use CoffeeScript for .coffee assets and views
-# gem 'coffee-rails', '~> 4.1.0'
-# See https://github.com/rails/execjs#readme for more supported runtimes
-# gem 'therubyracer', platforms: :ruby
-
-# Use jquery as the JavaScript library
-gem 'jquery-rails'
-# Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
-gem 'jbuilder', '~> 2.0'
-# bundle exec rake doc:rails generates the API under doc/api.
-gem 'sdoc', '~> 0.4.0', group: :doc
-
-gem 'haml'
-
-gem 'simple_form'
-
+gem 'active_skin'
+gem 'activeadmin'
+gem 'activeadmin_addons'
+gem 'acts-as-taggable-on', '~> 6.0'
+gem 'bootstrap-sass', '~> 3.4.1'
 gem 'devise'
-
-gem "paperclip", "~> 5.0.0"
-
-gem "cocoon"
-gem 'owlcarousel-rails'
-gem 'acts-as-taggable-on', '~> 3.4'
-gem 'rails3-jquery-autocomplete'
-gem "font-awesome-rails"
-gem "select2-rails"
+gem 'font-awesome-rails'
+gem 'jbuilder', '~> 2.0'
+gem 'jquery-rails'
+gem 'jquery-ui-rails', '~> 4.2.1'
+gem 'paperclip', '~> 5.0.0'
+gem 'pg'
+gem 'sass-rails', '~> 5.0'
+gem 'sdoc', '~> 0.4.0', group: :doc
+gem 'simple_form'
+gem 'turbolinks', '~> 5.2.0'
+gem 'uglifier', '>= 1.3.0'
+gem 'webpacker', '~> 5.x'
 gem 'will_paginate', '~> 3.1.0'
-gem 'gmaps4rails'
+
 # Use ActiveModel has_secure_password
 # gem 'bcrypt', '~> 3.1.7'
 
@@ -49,7 +35,9 @@ gem 'gmaps4rails'
 
 group :development, :test do
   # Call 'byebug' anywhere in the code to stop execution and get a debugger console
-  gem 'byebug'
+  gem 'annotate'
+  gem 'pry-byebug'
+  gem 'pry-rails'
   gem 'better_errors'
 end
 
@@ -63,7 +51,6 @@ group :development do
   gem 'capistrano', '3.6.0'
   gem 'capistrano-bundler', '~> 1.1', '>= 1.1.4'
   gem 'capistrano-rails', '~> 1.1', '>= 1.1.7'
-  gem 'capistrano-rbenv', github: "capistrano/rbenv"
+  gem 'capistrano-rbenv', github: 'capistrano/rbenv'
+  gem 'capistrano-webpacker-precompile', require: false
 end
-
-gem "cancan"
