@@ -2,13 +2,9 @@ class ApplicationController < ActionController::Base
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
-  before_filter do
+  before_action do
     resource = controller_name.singularize.to_sym
     method = "#{resource}_params"
     params[resource] &&= send(method) if respond_to?(method, true)
-  end
-  
-  def get_autocomplete_items(parameters)
-    super(parameters).presence || [OpenStruct.new(id: '', parameters[:method].to_s => '')]
   end
 end
