@@ -1,26 +1,20 @@
 class PagesController < ApplicationController
+  def buy; end
 
-  def buy
+  def delivery; end
 
-  end
+  def flavours; end
 
-  def flavours
-  end
+  def home; end
+
+  def location; end
 
   def press
+    @press_items = Press.all.order(name: :asc).paginate(page: params[:page], per_page: 24)
+  end
+  def companies
+    @companies = Company.all.order(name: :asc).paginate(page: params[:page], per_page: 24)
   end
 
-  def location
-  end
-
-  def us
-  end
-
-  def delivery
-  end
-
-  def mail_test
-    DeliveryMailer.sample_email().deliver
-  end
-
+  def us; end
 end
