@@ -1,6 +1,3 @@
-# config valid only for Capistrano 3.1
-lock '3.6.0'
-
 set :application, 'Artorta'
 set :repo_url, 'git@github.com:oaestay/Artorta.git'
 
@@ -23,7 +20,7 @@ set :deploy_to, '/home/deploy/Artorta'
 # set :pty, true
 
 # Default value for :linked_files is []
-set :linked_files, %w{config/database.yml config/secrets.yml}
+# set :linked_files, %w{config/database.yml config/secrets.yml}
 
 # Default value for linked_dirs is []
 set :linked_dirs, %w{bin log tmp/pids tmp/cache tmp/sockets vendor/bundle public/system}
@@ -46,5 +43,5 @@ namespace :deploy do
 
   after :publishing, :restart
   after :finishing, 'deploy:cleanup'
-
+  after 'deploy:updated', 'webpacker:precompile'
 end
